Updated releng/autorelease Project: yangtools  13bb06fa8a386783fa4796a8527b033b6d0f0271
BUG-4145: check if entire YangInstanceIdentifier is present in SchemaContext

A failure to find the SchemaNode for a path argument results in a
NullPointerException. We need the SchemaNode to understand whether it
targets an augmentation, as those need to be pruned from string output.

This patch improves the behaviour by throwing an
IllegalArgumentException with a description of which component we failed
to find.

Change-Id: I128eb63cf05b278b1af7516d426a6f0bafa87b90
Signed-off-by: Robert Varga <rovarga@cisco.com>