Updated releng/autorelease Project: controller  5c900a0f949b20d13e91a5464ccc292415b90cc5
Bug 4317: Added the version of yang-maven-plugin.

In the pom.xml of config-parent, the version of yang-maven-plugin was
missing. In the result, Maven downloaded the latest release version
(0.7.1-Lithium-SR1) of yang-maven-plugin, although the latest SNAPSHOT
version (0.8.0-SNAPSHOT) was expected here.
This patch added the version of yang-maven-plugin, and specified the
latest SNAPSHOT version for it.

Change-Id: I9a30aa4fc3c80d81c9342664135be484a219493d
Signed-off-by: Hideyuki Tai <Hideyuki.Tai@necam.com>