Updated releng/autorelease Project: integration/distribution  e8f2ea508f29ee74276b208c4b0804113bd6b605
Update features to fix build failures.

This patch introduces two changes which are needed to fix build
failures.

1. Update feature name of Lisp Flow Mapping project.

The Lisp Flow Mapping project has changed the name of
"odl-lispflowmapping-all" feature. The new name of the feature is
"odl-lispflowmapping-msmr".

2. Remove legacy features of OpenFlow Plugin project.

The OpenFlow Plugin project has deleted the features-legacy, and it is
not part of Beryllium.

Change-Id: Id00be5936ee296e274f518bef6b1f991aa06978e
Signed-off-by: Hideyuki Tai <Hideyuki.Tai@necam.com>