Updated releng/autorelease Project: vpnservice  078ae023c9cceb14aaadea10c81a5f1d90f47789
Use features-restconf from netconf project repository

This edit was omitted from Change 26456.

The previous controller feature repository is still present in Nexus,
but it should not be used, and it is not defined as dependency.

Since Change 26348, this omission is causing failures
in merge and distribution jobs across all ODL projects,
because SingleFeatureTest tries and fails to see the controller repository.

Change-Id: I23a4e326eaa280c659226a0a2b20c13c72d67e0b
Signed-off-by: Vratko Polak <vrpolak@cisco.com>