Updated releng/autorelease Project: ovsdb  fb5262e5de6bc81f2b9a3b8cdb9419f497ca2402
Bug 3989 -- Fix the original patch. Original patch can break
the functioning if there are more then one data change and
the first data change is OvsdbNodeAugmentation for which connection
already exist. It will return out of OnDataChanged() when it see
that connection already exist, without processing rest of the
data

Change-Id: Icec9fdcc241fb9f1b823d73cbe2dfdd1f5c3def7
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>