Updated releng/autorelease Project: ovsdb  908bc9051d15c7e1cad38b0022eba30ee0b513a1
Controller initiated connections always connect to one manager
listening in ptcp: mode. In clustered mode, all the three controller
will connect to the device, and in case any of the connection will flip
that instance will remove OvsdbNode from operational data store because
it will see only one manager in manager list. But because switch is
listening in passive mode, there can be multiple connection to same manager.
To handle this scenario, this patch adds number_of_connection attribute
to OvsdbNodeAugmentation to maintain the number of active connection to
specific manager. In case of switch initiated connection this value is
going to be 1 for each manager, but in case switch listening in passive
mode, this value will represent active connections (extraced from
Manager tables status column n_connections attributes).

Change-Id: I910104a5109b5bed6af7786a7692c617bd4305b7
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>