Updated releng/autorelease Project: mdsal  c4094ebe466c60eefed093cce40c080e4bb11461
Add common entity ownership interfaces

Added genericized entity ownership interfaces that are common between
the DOM and binding interfaces.

The instance identifier type in Entity is what differs between DOM and binding
but it's referenced in most the classes. For the classes that were
genericized, I prefixed then with "Generic" for lack of another/better
name. The intent is to keep the previous class names for the binding
versions (eg Entity, EntityOwnershipService etc), as they will be used most
commonly, and prefix the DOM versions with "DOM" to be consistent with other
md-sal APIs (eg DataBroker). The common classes for the data broker are
prefixed with "Async" but that doesn't really mean anything here so I went
with "Generic".

The other option is to keep the previous class names for the common API
and prefix the derived classes with "Binding" and "DOM" appropriately.
But we've tried to hide the binding concept as mauch as possible so
prefixing with "Binding" would add noise.

Change-Id: I00013f61f7c848c99fdc30740c8ab323448cc976
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>