Updated releng/autorelease Project: aaa  b50ad56f527d6da5b14cec06d2c6c1396feec9a4
Support enabling and disabling shiro based AAA

This change supports enabling and disabling shiro AAA through the use of a
ServiceProxy.  This support was added in order to support workflows such as
those included in the odl-restconf-noauth feature.  This change also adds a
significant amount of documentation to help users extend and debug the shiro
functionality.

Change-Id: I213a200c18d48fd1bf59da1ec171b0814a283fce
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>