Updated releng/autorelease Project: mdsal  967aafaff5cdf01e86fcacde893b7e30e5c3ca7f
Wait for schemas when transforming BA->NormalizedNode

If the transformation fails due to missing schema, codec tries to wait for
the models to become available. The assumption is that if there already are
the BA classes, its likely that the schemas will follow shortly.

This improvement should fix possible race conditions during ODL startup.

Change-Id: I101b77d6ca8f247c271bed0a18fff20de43c24bf
Signed-off-by: Maros Marsalek <mmarsale@cisco.com>