Updated releng/autorelease Project: controller  7a2d5cb49eab242e10233e0218ee1906c13df901
Revert ModuleInfoBundleTracker patches

Reverted patch https://git.opendaylight.org/gerrit/#/c/27138/ as it
causes some feature tests to take a long time due to the busy wait.
Also it appears the ModuleFactory OSGi services are needed as the
BlankTransactionServiceTracker listens for them (I'm not clear what this
does). I'll try to figure out another way to accomplish the intent
of the reverted patch.

Change-Id: Ifc91dada86ac7feee1a0a9390a55e68d7f113153
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>