Updated releng/autorelease Project: ovsdb  83cfd5bd73f79b3b10e71a13b8f6563ea37fee20
Use the memorized service reference

Currently when dispatching "event handler added" events we retrieve
the appropriate service reference from the bundle context, but this
seems to cause problems when multiple services are registered for the
same interfaces. Since we know the service reference we got when
registering the service involved, we can just pass that.

As a result, ConfigInterface::setDependencies() doesn't need a
BundleContext at all, and since we're no longer supposed to pass that
around, we can just remove it altogether.

Some more debugging information should help track down other issues of
this type.

Change-Id: Iac54595aed388165711a7c4a3d86577646c59d4f
Signed-off-by: Stephen Kitt <skitt@redhat.com>