Updated releng/autorelease Project: bgpcep  30141a43a11bd8836ea70211b254f2bfda41863b
BUG 4413: Unit test "BGPDispatcherImplTest.testCreateClient" fails

Test testCreateClient/reconnectClient sometimes fails
because port still not free from previous tests.
Since NeverReconnectStrategy connections strategy is used,
when first attempt to stablish connection fails,
it throws throwable. As a solution we set the Socket on test
to be able to be reused.

Change-Id: I55d56ae2fbbf402924ba86a447ff518c329f5410
Signed-off-by: Claudio D. Gasparini <cgaspari@cisco.com>