Updated releng/autorelease Project: vtn  acd0e860dc5420c3a068fc8aaa2c3c65f08e1423
Updated UT code for VTN Manager Neutron.

This change fixes the compilation failure which is introduced by the
recent change in Neutron project.
https://git.opendaylight.org/gerrit/#/c/25500/
In the above patch, the Neutron project added public new methods to
NeutronObject class in the org.opendaylight.neutron.spi package, and
this caused the compilation failure.

Change-Id: I2214ccef8d21b6ec9b25ac4f82956417b449f47f
Signed-off-by: Hideyuki Tai <Hideyuki.Tai@necam.com>