Updated releng/autorelease Project: ovsdb  fd44e6a875c671a068162f9e8a013d6b6759b8c4
Fix issue that was causing integration test failure.

In case of controller initiated connection, iid stamping
was not happening. Before clustering code merge, iid stamping
was done in updateData(), but after clustering code merge,
it does not allow to make any changes to the device till
owner of the device is selected by EntityOnwershipService,
so updateData() was not stamping the iid to swtich. With this
fix, once owner of the device is selected, it explictly stamp
iid to the switch.

NOTE: we should revert following patch before merging this patch
https://git.opendaylight.org/gerrit/#/c/28309/1

I ran the integration suit locally and it passes all the tests

[main] INFO org.ops4j.pax.exam.spi.reactors.ReactorManager - suite finished
Tests run: 22, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 288.094 sec - in org.opendaylight.ovsdb.southbound.it.SouthboundIT

Results :

Tests run: 22, Failures: 0, Errors: 0, Skipped: 0

Change-Id: I35d6e0a4fa94eef019420e9b9603873d5088a934
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>