Updated releng/autorelease Project: bgpcep  bb604e5536c7a8e9d6a35fda5ff9952180fd2001
BUG-3778: Next Hop needs to be specified in each extension

Since each NextHop need to be specified for each extension,
we need to enforce users to register his corresponding handlers.
Thus a new registerNlriParser method is created with NexHop
handler and NextHop classes which are handled, as a new parameters.
Previous registerNlriParser method is deprecated.

Change-Id: I7aca6dbc679bf146e9333e7f2d2c9affb75808e8
Signed-off-by: Claudio D. Gasparini <cgaspari@cisco.com>