Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Good idea although I'm not a fan of the *.repo files as I feel it duplicates work. Can we not introspect using the pom.xml file that exists in the repo?

I would hope that someone could "git checkout release/lithium-sr2" and run ./scripts/releasenotes/generate-release-notes.sh and it will run the script and be able to get all the information for the current release with minimal parameters passed.

We could make the script accept additional optional parameters to tweak if they don't want to run against the current checked out commit. I think the script will be more elegant this way and we can backport it to stable/lithium if we want it there too. Thoughts?

Patch-set: 1
Label: Code-Review=-1