Update patch set 1
Patch Set 1:

Your comments are all valid, but my goal is to break this into three phases.  The first phase is an initial pass based on existing scripts with minimal changes (don't break what works).  In the second phase, we should refactor the scripts to 1) execute with minimal parameters, 2) autogenerate the list of projects depending on the branch (read from pom), 3) format the wiki and ascii outputs to be ready-to-push complete pages, and 4) executed by a Jenkins job.

Patch-set: 1