Updated releng/autorelease Project: ovsdb  f04bf19a5210aa2606cdd60b00d1522d75401103
Revert SouthboundIT multi-threading and refactor

Running the tests in ExecutorService-managed threads masks a number of
errors which leave the target OVSDB in an inconsistent state. This
patch reverts the multi-threading and subsequent refactor, while
preserving some of the later changes (notably the Jacoco activation).

The multi-threading and testcase-refactoring will be re-done as
follow-up patches.

Change-Id: I9d534509f772afd75877d230cf4e2dc4654fab85
Signed-off-by: Stephen Kitt <skitt@redhat.com>